+++ app/src/main/java/com/kontvip/wisecoin/domain/display/PaymentDisplay.kt
@@ -6,6 +6,6 @@ interface PaymentDisplay : UiState.UiDisplay {
     fun displayTime(time: Long)
     fun displayDescription(description: String)
     fun displayCategory(category: String)
-    fun displayAmount(amount: Int)
+    fun displayAmount(amount: Double)
     fun displayImage(image: String)
 }
