+++ app/src/main/java/com/kontvip/wisecoin/presentation/core/DomainToUiPaymentMapper.kt
@@ -5,7 +5,7 @@ import com.kontvip.wisecoin.presentation.model.PaymentUi
 
 class DomainToUiPaymentMapper : PaymentDomain.Mapper<PaymentUi> {
     override fun map(
-        id: String, time: Long, description: String, category: String, amount: Int, image: String
+        id: String, time: Long, description: String, category: String, amount: Double, image: String
     ): PaymentUi {
         return PaymentUi(id, time, description, category, amount, image)
     }
